somethid done and modified
