+++ src/Components/Navbar/Navbar.css
@@ -4,12 +4,14 @@
     display: flex;
     position: fixed;
     top: 0;
+    background-color: rgb(45, 45, 45);
     width: 100%;
     z-index: 100;
     justify-content: space-between;
     align-items: center;
     backdrop-filter: blur(10px);
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(1000,600, 1, 2);
+    /* box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(1000,600, 1, 2); */
+    box-shadow: 0 4px 8px 0 rgb(105, 14, 14);
     background: linear-gradient( white);
   }
