--- src/Components/ReuseableCard/ReuseableCard.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import './ReuseableCard.css'
-
-function ReuseableCard() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default ReuseableCard
