+++ src/Components/ReuseableCard/TeamCard/TeamCard.css
@@ -1,0 +1,52 @@
+.card {
+    width: 250px;
+    position: relative;
+    padding-top: 105px;
+    border: 2px solid #00ffdd;
+    border-radius: 15px 0 15px 0;
+    height: 180px;
+    margin: 30px 40px;
+    transition: all 0.5s ease-out;
+    box-shadow: 0 8px 12px 0 rgba(0, 0, 0, 0.2), 0 12px 30px 0 rgba(0, 0, 0, 0.19);
+    background-image: url("https://www.transparenttextures.com/patterns/dirty-old-black-shirt.png");
+  }
+  
+  .card:hover {
+    transform: scale(1.05);
+  }
+  
+  .card-image {
+    width: 180px;
+    height: 180px;
+    border-radius: 500px;
+    object-fit: cover;
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+    top: -50%;
+  }
+  
+  .card h1 {
+    font-size: 1.4rem;
+    font-weight: 500;
+    text-transform: uppercase;
+    letter-spacing: 2px;
+    text-align: center;
+    color: #00ffdd;
+  }
+  
+  .card h2 {
+    font-size: 1rem;
+    font-weight: 500;
+    text-transform: uppercase;
+    letter-spacing: 2px;
+    text-align: center;
+    margin-top: 4px;
+    color: rgba(0, 255, 221, 0.9);
+  }
+  
+  @media (max-width: 600px) {
+    .card {
+      margin: 60px 40px;
+    }
+  }
