+++ src/Components/SocialMedia/SocialMedia.css
@@ -1,0 +1,21 @@
+  .social-media{
+    display: flex;
+    float: left;
+    left: 0;
+  }
+  .MuiIconButton-root {
+    display: flex;
+    flex-direction: column;
+    /* float:left; */
+    /* margin-left: 0; */
+    /* margin-right: auto; */
+  }
+
+  .MuiIconButton-root:hover {
+    transform: scale(1.1);
+    color: #ff0000;
+  }
+  
+  .MuiSvgIcon-root {
+    font-size: 28px;
+  }
