+++ src/Pages/Home/Home.css
@@ -1,0 +1,24 @@
+  .home {
+    height: 100vh;
+    background-color: #302f2f;
+    overflow: hidden;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    display: flex; 
+   justify-content: center;
+   align-items: center;
+  }
+
+  .name-and-tagline{
+    /* align-items: center; */
+    margin-left: 20px;
+  }
+  .social-media{
+
+  }
+
+  
+
