+++ src/Pages/Team/Team.js
@@ -4,7 +4,7 @@ import './Team.css'
 function Team() {
   return (
     <div>
-      
+     
     </div>
   )
 }
